+++ new_code/writeCSV.cpp
@@ -11,7 +11,7 @@ using std::cout;
 
 void writeCSV(const char* filename, const float* data, size_t n_rows, size_t n_cols){
     FILE *fileOut = fopen(filename,"w"); // file name is passed as a function argument, open for write
-    int i; // loop index variables
+    size_t i; // loop index variables
 
     // loop through the input array, splitting into rows and columns according to the input sizes
     /*
